blob: fix math.MaxUint32 overflow on 32-bit platforms
Commit 0ce8e60f added a guard comparing `d.bd.Rows()` (which returns
`int`) against `math.MaxUint32` to prevent overflow in the subsequent
`uint32(d.bd.Rows()+1)` cast. On 32-bit platforms, `int` is 32 bits and
the untyped constant `math.MaxUint32` (4294967295) overflows `int`,
causing a compile error.

Cast to `uint32` before comparing. The round-trip `uint32 → int → uint32`
preserves the original header value on both 32-bit and 64-bit platforms.

Fixes #5858
Fixes #5859

+++ sstable/blob/blocks.go
@@ -163,7 +163,7 @@ func (d *indexBlockDecoder) Init(data []byte) {
 	d.bd.Init(data, indexBlockCustomHeaderSize)
 	d.virtualBlocks = colblk.DecodeColumn(&d.bd, indexBlockColumnVirtualBlocksIdx,
 		uint32(d.virtualBlockCount), colblk.DataTypeUint, colblk.DecodeUnsafeUints)
-	if d.bd.Rows() == math.MaxUint32 {
+	if uint32(d.bd.Rows()) == math.MaxUint32 {
 		panic(errors.AssertionFailedf("invalid row count"))
 	}
 	// Decode the offsets column. We pass rows+1 because an index block encoding
